Add raid note adjustment text.
This update allows players to put an adjustment "note" between : : and
have an adjustment automatically created for the user.  The created
adjustment is ALWAYS of 0 WRS value and is used for a placeholder to
identify a specific raid standing.

This should help answer the question "why did I not get full WR credit for
last night's raid?"